fix(ffi): pin every libdeluge enum to an explicit fixed width
arm-none-eabi-gcc defaults to -fshort-enums, so the shipping device C++
lays out every libdeluge FFI enum as 1 byte. A prior commit removed
-fshort-enums from bindgen but pinned only DelugeRegionState explicitly,
so the device bindgen target started emitting int-sized enums while
arm-gcc kept emitting short ones -- DelugeInputEvent mismatched between
Rust (8 bytes) and the arm C++ (6 bytes), corrupting pad-press/note-off
decoding. The sim golden runs on the host, where enums were already
int-sized on both sides, so it never caught this.

Give the other 10 enums the same explicit fixed-underlying-type
treatment DelugeRegionState already had, each pinned to its current
arm-short width (uint8_t, DelugeStatus as int8_t for its negative error
codes). An explicit underlying type is authoritative in C and C++
regardless of -fshort-enums, so bindgen now emits the same 1-byte width
on every target -- device, host_app, and host stand-ins alike -- and
the shipping device layout is unchanged.

Revert the Rust-side widenings the prior fix made now that every enum
is narrow everywhere: control.rs's DelugeInputEvent/Kind size asserts
(8/4 -> 6/1), sys_host.rs's hand-written DelugeStatus/DelugeInputEventKind
stand-ins and DelugeInputEvent layout, and sd.rs's DelugeStatus/
DelugeCardEvent host stand-ins. Reworded comments in build.rs,
lens1_vt_sim/build.rs, HOST_HARNESS.md, and abi.rs that attributed enum
width to -fshort-enums or an int/short compiler default -- the width is
now pinned explicitly in each header, independent of any flag.

+++ src/bsp/rust/src/sd.rs
@@ -75,14 +75,14 @@ use crate::sys::{
 // its top-of-file comment), so `crate::sys` simply doesn't exist to import on
 // host, independent of what deluge-sdk exposes. Mirror the C types' shapes
 // directly here instead (same convention as fiber.rs's `RunCondition`):
-// `DelugeStatus` is `i32` (its values run -13..=0, forcing a signed type;
-// plain int-sized C enums throughout the build — no `-fshort-enums`, see
-// `../build.rs`), `DelugeCardEvent` is `u32` (0..=2, non-negative so clang
-// picks `unsigned int`). Values match include/libdeluge/{types,block_device}.h
-// exactly, so the full card-event edge-tracking algorithm in
+// `DelugeStatus` is `i8` (its values run -13..=0, forcing a signed type; the
+// header pins its underlying type explicitly — `enum DelugeStatus : int8_t`),
+// `DelugeCardEvent` is `u8` (0..=2, pinned explicitly as `enum DelugeCardEvent
+// : uint8_t`). Values match include/libdeluge/{types,block_device}.h exactly,
+// so the full card-event edge-tracking algorithm in
 // `deluge_block_poll_card_event` below runs identically on both targets.
 #[cfg(not(target_os = "none"))]
-type DelugeStatus = i32;
+type DelugeStatus = i8;
 #[cfg(not(target_os = "none"))]
 const DELUGE_OK: DelugeStatus = 0;
 #[cfg(not(target_os = "none"))]
@@ -93,7 +93,7 @@ const DELUGE_ERR_NODEV: DelugeStatus = -6;
 const DELUGE_ERR_WRITE_PROTECTED: DelugeStatus = -12;
 
 #[cfg(not(target_os = "none"))]
-type DelugeCardEvent = u32;
+type DelugeCardEvent = u8;
 #[cfg(not(target_os = "none"))]
 const CARD_NONE: DelugeCardEvent = 0;
 #[cfg(not(target_os = "none"))]
